fix terminal spawn shell handling on Windows
